Handle paths hover with custom hook
